Code reorganisation

+++ ALEX/src/robot/Joint.h
@@ -13,7 +13,6 @@
 #include <sys/un.h>
 #include <sys/socket.h>
 
-
 //Node ID for the 4 joints
 
 #define LEFT_HIP 1
@@ -22,7 +21,6 @@
 #define RIGHT_KNEE 4
 #define LEFT_ANKLE 5
 #define RIGHT_ANKLE 6
-
 
 using namespace std;
 
@@ -34,11 +32,11 @@ class Joint
     double qd;
     long maxq, minq;
     long maxdq, mindq;
-	double converterA, converterB;
+    double converterA, converterB;
 
     void setPos(long qd);
     void setVel(long dqd);
-    void motorPosConverter(double origDeg, long * newMotorCmnd, int nodeid);
+    void motorPosConverter(double origDeg, long *newMotorCmnd, int nodeid);
     double motorPosToDegConverter(long motorCmdAngle, int nodeid);
     void calcAB(long y1, long x1, long y2, long x2, double *A, double *B);
     int bitFlipState;
@@ -56,7 +54,7 @@ public:
     void applyPos(long qd);
     void applyPosDeg(double qd);
     void applyVel(long dqd);
-    
+
     void disable();
     void readyToSwitchOn();
     void enable();
@@ -64,14 +62,14 @@ public:
     int getPos();
     double getPosDeg();
     double getDesPosDeg();
-    
+
     void printInfo();
     void updateJoint();
-   void testWrite();
+    void testWrite();
     bool bitflipHigh();
     bool bitflipLow();
     int getStatus();
-    double getActualTorque();
+    int getActualTorque();
     int getBitFlipState();
     void setBitFlipState(int bit);
